Vim settings and timeouts in tmux config
Pressing ESC to return to normal mode in Vim has about a 1 second delay.
To fix it, add a timeout value of 0 to the TMUX config, and set a low
ttimeout value inside Vim.

In non-tmux terminals it may also be necessary to set KEYTIMEOUT=1
(value is timeout in milliseconds)